requirements.txt
fastapi==0.88.0
uvicorn==0.18.3
python-telegram-bot==20.0
requests==2.26.0
python-dotenv==0.21.1
pydantic==1.10.2
tortoise-orm==0.17.7
databases==0.5.5
alembic==1.7.7
aiohttp==3.7.4.post0
pytest==7.1.2
black==22.3.0
autopep8==1.5.7
mypy==0.931
typer==0.4.0
apscheduler